pyproject.toml
[project]
name = "yt-notetaker"
version = "0.1.0"
description = "Generate DOCX notes from YouTube playlists or videos with optional LLM summaries"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "yt-dlp>=2024.4.9",
  "python-docx>=1.1.0",
  "requests>=2.31.0",
  "typer>=0.12.0",
  "streamlit>=1.33.0",
  "langchain-openai>=0.2.0",
  "langchain>=0.2.0",
  "pydantic>=2.5.0",
]

[project.scripts]
yt-notetaker = "main:main"

[build-system]
requires = ["hatchling>=1.25.0"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["yt_notetaker"]
